scripts を型検査の対象に入れる（引数漏れが実行時まで気づけなかった）
measure_topix100.py が list_symbols() missing 1 required positional
argument: 'as_of' で落ちた。mypy の files が ["src", "tests"] で
scripts を含んでいなかったのが原因。

さらに autotrader パッケージに py.typed がなかったため、インストール
済みパッケージとして import した際に中身が解析されず、仮に scripts を
対象にしても同じ誤りを見逃していた。py.typed を追加して package-data
で同梱し、files に scripts を追加した。

この変更だけで他に4件の潜在バグが見つかり、あわせて修正した:
- analyze_expectancy.py: wilson_interval に float を渡していた
  （同じ関数の別の呼び出しでは int() で包んでいたのに1箇所だけ漏れ）
- verify_data_sources.py: date | None を date 引数に渡していた
  （has_range で守られてはいるがプロパティ越しで型検査は追えない）
- float ** float が Any に落ちる箇所が2件

手元で回す実測スクリプトだからと検査対象から外していたが、実測
スクリプトこそ結果の正しさに直結する。

list_symbols が as_of を必須にしているのは意図的な設計（現在の銘柄
一覧を過去に適用するとサバイバーシップバイアスになる）なので、
引数を足す形で修正した。契約範囲は400応答から自動学習されるので、
範囲外を叩いて学習させてから終端へ寄せる。

docs/00 意思決定ログ62に記録した。

+++ scripts/analyze_expectancy.py
@@ -78,7 +78,7 @@ def required_win_rate(edge_per_trade_atr: float) -> float:
 def edge_for_monthly_target(monthly_return: float) -> float:
     """月利目標から1トレードあたりに要る優位（ATR単位）を逆算する。"""
     trades_per_month = TRADES_PER_DAY * BUSINESS_DAYS_PER_MONTH
-    per_trade_equity = (1.0 + monthly_return) ** (1.0 / trades_per_month) - 1.0
+    per_trade_equity = float((1.0 + monthly_return) ** (1.0 / trades_per_month)) - 1.0
     # 建玉は資金の 25%。資金に対する利益率を建玉に対する値幅へ戻す
     per_trade_position = per_trade_equity / MAX_WEIGHT_PER_SYMBOL
     return per_trade_position / MEDIAN_ATR_PCT
@@ -86,7 +86,7 @@ def edge_for_monthly_target(monthly_return: float) -> float:
 
 def monthly_from_period(total_return: float, days: int) -> float:
     """期間リターンを月次（20営業日）に換算する。"""
-    return (1.0 + total_return) ** (BUSINESS_DAYS_PER_MONTH / days) - 1.0
+    return float((1.0 + total_return) ** (BUSINESS_DAYS_PER_MONTH / days)) - 1.0
 
 
 CAPITAL = 500_000.0
@@ -123,7 +123,9 @@ def main() -> None:
     )
 
     measured = MEASURED["ブレーカー無効"]
-    lo, hi = wilson_interval(measured["win_rate"] * measured["trades"], measured["trades"])
+    lo, hi = wilson_interval(
+        measured["win_rate"] * measured["trades"], int(measured["trades"])
+    )
     verdict = "外側（構造的に負けている）" if hi < breakeven else "内側（判定不能）"
     print(f"  実測 {measured['win_rate']:.1%} の信頼区間 [{lo:.1%}, {hi:.1%}] は分岐点の{verdict}")
